Update (#1)
* Fix missing v4 suffix

* Fix readme typo

* Fix soft delete issues

* Add missing features to readme

* Update changelog, bump version, fix bindata

* Support `oid` type

* Fix for the extremely slow model generation with some MySQL databases.

If one tries to generate the models on a MySQL database with a lot of
databases and tables the generation is getting very slow (#329).

This seems to result from MySQL not optimizing queries on views very well.

The solution I have chosen and validated is to replace the table and schema
in the subquery with the actual names instead of references to the parent
queries.

The result for my system is insane and went down from over 9 hours to 4
seconds.

This fix is kinda ugly but straight forward and I think the speed advantage
is worth it. Especially as the code is never being seen by anyone.

* MariaDB Information Schema differs from MySQL

These changes need to be made to have a similar result for default
values in sqlboiler when using MariaDB instead of MySQL

* added relation-tag flag & config option

* Bump version

* Fix changelog missing date

* Added missing comma from last commit!

* Fix spacing issues

* Update changelog, bump version

* WIP

* WIP

* Fix mssql tests

* Fix broken test due to go modules

* Allow targeting column by table name in types.match

Looks for "tables" in types.match and if present,
will replace the columns only on those tables.

* Fix $1 in mysql soft delete generation

* Bump version

* Update README.md

* Add "alias" struct tag casing

We have number of options for struct tag casing, but we have option that
respect our choice in alias section of configurations.

* Update README.md

* Ignore generated columns for psql

psql's generated columns cannot be set. So this PR add condition to filter those columns.
sqlboiler's mysql driver already ignores virtual columns (which is generated column in psql) in #266

* Add types.DecimalContext

* In MySQL driver increased control over ssl-mode to work around volatiletech/sqlboiler#273

* Update changelog

* Fix where not in bug (empty lists)

* Update documentation

* Update changelog

* Update changelog

* Fix mistake in comments

* Additional fix for mysql soft deletion

* Update readme

* Fix godoc link

* Document deleted_at/updated_at interaction

* Add note about autogenerated tests failing

* Make binding cache accesses allocation-free

* Fix go-bindata from bad merge

* Fix boilingcore tests using published v4 module

* Update changelog

* Update changelog

* Update changelog and version

* Fix changelog

Co-authored-by: Aaron L <aaron@bettercoder.net>
Co-authored-by: V-X <v@nul.lu>
Co-authored-by: Hans Raaf <hara@oderwat.de>
Co-authored-by: slt <slt@sp-online.de>
Co-authored-by: Stephen Afam-Osemene <stephen.afam-osemene@engelvoelkers.com>
Co-authored-by: Patrick O'Brien <patrick@volatile.tech>
Co-authored-by: Darkclainer <Darkclainer@gmail.com>
Co-authored-by: Cho Chi Him <chochihim@gmail.com>
Co-authored-by: MattMc <4984708+tooolbox@users.noreply.github.com>
Co-authored-by: mfzy <49176317+mfzy602@users.noreply.github.com>
Co-authored-by: zikaeroh <48577114+zikaeroh@users.noreply.github.com>